Struct potpourri::backend::ndarray::gaussian::Gaussian

pub struct Gaussian {
    pub means: Array2<f64>,
    pub covariances: Array3<f64>,
    pub precisions: Array3<f64>,
    /* private fields */
}

Represents a gaussian mixture model. The number of components is determined automatically from the responsibilities when first calling maximize method.

The sufficient statistics that can be extracted from the data to maximize the parameters. It is a triple of arrays (names as in Kimura et al.):

$$ \begin{aligned} a_j &= \sum_i^n r_{ij},&& (k) \\ b_j &= \sum_i^n r_{ij} \cdot x_i ,&& (k \times d) \\ c_j &= \sum_i^n r_{ij} \cdot x_i^T\cdot x_i, &&(k \times d \times d) \\ \end{aligned} $$

Fields
means: Array2<f64>

The mean values, $ k\times d $

covariances: Array3<f64>

The covariance matrices), $(k\times d\times d)$

precisions: Array3<f64>

The precision matrices (inverted coariances), $(k\times d\times d)$

Implementations

impl Gaussian

pub fn new() -> Gaussian